% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarized_rebase.R
\name{solarized_rebase}
\alias{solarized_rebase}
\title{stolen from ggthemes; not exported}
\usage{
solarized_rebase(light = TRUE)
}
\arguments{
\item{light}{T/F switch between solarized light (gold background) or dark (navy bg)}
}
\description{
stolen from ggthemes; not exported
}
