% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{ipsum_palette}
\alias{ipsum_palette}
\title{A muted, qualitative color palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
ipsum_palette
}
\description{
A muted, qualitative color palette
}
\keyword{datasets}
