% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_solarized_mk.R
\name{theme_solarized_mk}
\alias{theme_solarized_mk}
\title{custom GGPLOT2 theme for insertion
adapted from ggthemes()}
\usage{
theme_solarized_mk(base_size = 12, base_family = "", light = TRUE)
}
\description{
custom GGPLOT2 theme for insertion
adapted from ggthemes()
}
